refactor: moved overpass handling into own directory
- added markerHandler

Refs: #7
